ST6RI-884 Moved EvaluationUtil into org.omg.sysml.util with other Utils.
