+++ org.omg.sysml.execution/src/org/omg/sysml/execution/expressions/functions/SumFunction.java
@@ -23,11 +23,11 @@ package org.omg.sysml.execution.expressions.functions;
 
 import org.eclipse.emf.common.util.EList;
 import org.omg.sysml.expressions.ModelLevelExpressionEvaluator;
-import org.omg.sysml.expressions.util.EvaluationUtil;
 import org.omg.sysml.lang.sysml.Element;
 import org.omg.sysml.lang.sysml.InvocationExpression;
 import org.omg.sysml.lang.sysml.LiteralInteger;
 import org.omg.sysml.lang.sysml.LiteralRational;
+import org.omg.sysml.util.EvaluationUtil;
 
 public class SumFunction extends NumericalFunction {
