+++ org.omg.sysml.plantuml/META-INF/MANIFEST.MF
@@ -15,7 +15,6 @@ Import-Package: com.google.common.collect,
  org.eclipse.emf.ecore.util,
  org.omg.sysml.execution.expressions,
  org.omg.sysml.expressions,
- org.omg.sysml.expressions.util,
  org.omg.sysml.lang.sysml,
  org.omg.sysml.lang.sysml.impl,
  org.omg.sysml.lang.sysml.util,
