+++ org.omg.sysml/META-INF/MANIFEST.MF
@@ -32,7 +32,6 @@ Export-Package: org.omg.sysml,
  org.omg.sysml.delegate.setting,
  org.omg.sysml.expressions,
  org.omg.sysml.expressions.functions,
- org.omg.sysml.expressions.util,
  org.omg.sysml.lang.sysml,
  org.omg.sysml.lang.sysml.impl,
  org.omg.sysml.lang.sysml.util,
