+++ org.omg.sysml/src/org/omg/sysml/expressions/functions/ConditionalLogicalFunction.java
@@ -22,9 +22,9 @@ package org.omg.sysml.expressions.functions;
 
 import org.eclipse.emf.common.util.EList;
 import org.omg.sysml.expressions.ModelLevelExpressionEvaluator;
-import org.omg.sysml.expressions.util.EvaluationUtil;
 import org.omg.sysml.lang.sysml.Element;
 import org.omg.sysml.lang.sysml.InvocationExpression;
+import org.omg.sysml.util.EvaluationUtil;
 
 public abstract class ConditionalLogicalFunction extends ControlFunction {
