+++ org.omg.sysml/src/org/omg/sysml/expressions/functions/ControlFunction.java
@@ -26,10 +26,10 @@ import java.util.function.BiFunction;
 import org.eclipse.emf.common.util.BasicEList;
 import org.eclipse.emf.common.util.EList;
 import org.omg.sysml.expressions.ModelLevelExpressionEvaluator;
-import org.omg.sysml.expressions.util.EvaluationUtil;
 import org.omg.sysml.lang.sysml.Element;
 import org.omg.sysml.lang.sysml.Expression;
 import org.omg.sysml.lang.sysml.InvocationExpression;
+import org.omg.sysml.util.EvaluationUtil;
 
 public abstract class ControlFunction implements LibraryFunction {
