+++ org.omg.sysml/src/org/omg/sysml/expressions/functions/IsTypeFunction.java
@@ -22,10 +22,10 @@ package org.omg.sysml.expressions.functions;
 
 import org.eclipse.emf.common.util.EList;
 import org.omg.sysml.expressions.ModelLevelExpressionEvaluator;
-import org.omg.sysml.expressions.util.EvaluationUtil;
 import org.omg.sysml.lang.sysml.Element;
 import org.omg.sysml.lang.sysml.InvocationExpression;
 import org.omg.sysml.lang.sysml.Type;
+import org.omg.sysml.util.EvaluationUtil;
 
 public class IsTypeFunction extends BaseFunction {
