+++ org.omg.sysml/src/org/omg/sysml/expressions/functions/ListConcatFunction.java
@@ -24,9 +24,9 @@ package org.omg.sysml.expressions.functions;
 import org.eclipse.emf.common.util.BasicEList;
 import org.eclipse.emf.common.util.EList;
 import org.omg.sysml.expressions.ModelLevelExpressionEvaluator;
-import org.omg.sysml.expressions.util.EvaluationUtil;
 import org.omg.sysml.lang.sysml.Element;
 import org.omg.sysml.lang.sysml.InvocationExpression;
+import org.omg.sysml.util.EvaluationUtil;
 
 public class ListConcatFunction extends BaseFunction {
