+++ org.omg.sysml/src/org/omg/sysml/expressions/functions/TimesFunction.java
@@ -22,8 +22,8 @@
 package org.omg.sysml.expressions.functions;
 
 import org.eclipse.emf.common.util.EList;
-import org.omg.sysml.expressions.util.EvaluationUtil;
 import org.omg.sysml.lang.sysml.Element;
+import org.omg.sysml.util.EvaluationUtil;
 
 public class TimesFunction extends ArithmeticFunction {
